src/test/suite/extension.test.ts
import * as assert from 'assert';
import * as vscode from 'vscode';
// Uncomment the following line and adjust the path to your main extension file
// import * as myExtension from '../../extension';

suite('Extension Test Suite', () => {
  vscode.window.showInformationMessage('Start all tests.');

  // Example test for a command execution
  test('Command Execution Test', async () => {
    // Replace 'extension.generateTsConfig' with your command ID
    const commandResult = await vscode.commands.executeCommand(
      'extension.generateTsConfig',
    );

    // Assert based on expected behavior of the command (adjust accordingly)
    assert.strictEqual(
      commandResult,
      undefined,
      'Command did not return the expected result.',
    );
  });

  // Sample array index tests
  test('Array Index Test', () => {
    assert.strictEqual(-1, [1, 2, 3].indexOf(5));
    assert.strictEqual(-1, [1, 2, 3].indexOf(0));
  });
});
